+++ src/exactly_lib/instructions/multi_phase/new_dir.py
@@ -14,7 +14,7 @@ from exactly_lib.definitions.entity import syntax_elements
 from exactly_lib.instructions.multi_phase.utils import instruction_embryo as embryo
 from exactly_lib.instructions.multi_phase.utils import instruction_part_utils
 from exactly_lib.instructions.multi_phase.utils.assert_phase_info import IsAHelperIfInAssertPhase
-from exactly_lib.section_document.element_parsers.token_stream import TokenStream
+from exactly_lib.section_document.element_parsers.token_stream_parser import TokenParser
 from exactly_lib.symbol.data.path_sdv import PathSdv
 from exactly_lib.symbol.path_resolving_environment import PathResolvingEnvironmentPostSds
 from exactly_lib.symbol.symbol_usage import SymbolUsage
@@ -22,8 +22,8 @@ from exactly_lib.test_case.os_services import OsServices
 from exactly_lib.test_case.phases.common import InstructionEnvironmentForPostSdsStep, PhaseLoggingPaths
 from exactly_lib.test_case_utils.documentation import relative_path_options_documentation as rel_path_doc
 from exactly_lib.test_case_utils.err_msg import path_err_msgs
+from exactly_lib.test_case_utils.parse import parse_path
 from exactly_lib.test_case_utils.parse.rel_opts_configuration import argument_configuration_for_file_creation
-from exactly_lib.test_case_utils.parse.token_parser_extra import TokenParserExtra
 from exactly_lib.util.textformat.structure import structures as docs
 from exactly_lib.util.textformat.structure.core import ParagraphItem
 
@@ -105,13 +105,11 @@ class TheInstructionEmbryo(embryo.InstructionEmbryo[Optional[TextRenderer]]):
         return None
 
 
-class EmbryoParser(embryo.InstructionEmbryoParserThatConsumesCurrentLine):
-    def _parse(self, rest_of_line: str) -> TheInstructionEmbryo:
-        tokens = TokenParserExtra(TokenStream(rest_of_line))
-
-        target_path = tokens.consume_path(RELATIVITY_VARIANTS)
-        tokens.report_superfluous_arguments_if_not_at_eol()
-        return TheInstructionEmbryo(target_path)
+class EmbryoParser(embryo.InstructionEmbryoParserFromTokensWoFileSystemLocationInfo[Optional[TextRenderer]]):
+    def _parse_from_tokens(self, token_parser: TokenParser) -> TheInstructionEmbryo:
+        path = parse_path.parse_path_from_token_parser(RELATIVITY_VARIANTS, token_parser)
+        token_parser.report_superfluous_arguments_if_not_at_eol()
+        return TheInstructionEmbryo(path)
 
 
 PARTS_PARSER = instruction_part_utils.PartsParserFromEmbryoParser(